# apps/api/openapi-hunter.yaml
openapi: 3.0.3
info:
  title: In–Midst–My–Life Hunter Protocol API
  description: |
    Job search automation and career advancement engine.

    ## Overview
    The Hunter Protocol enables autonomous job search with intelligent resume
    and cover letter tailoring. Features are tiered and quota-based:

    - **FREE Tier**: 5 job searches/month, no document generation
    - **PRO Tier**: Unlimited searches, 5 resume tailoring, 3 cover letters
    - **ENTERPRISE Tier**: Unlimited everything

    ## Authentication
    All endpoints require bearer token in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Rate Limiting
    Quota-based rate limiting (per feature, per user):
    - Job searches: 5/month (FREE), unlimited (PRO/ENTERPRISE)
    - Resume tailoring: 0/month (FREE), 5/month (PRO), unlimited (ENTERPRISE)
    - Cover letters: 0/month (FREE), 3/month (PRO), unlimited (ENTERPRISE)
    - Auto-apply: 0/month (FREE), 5/month (PRO), unlimited (ENTERPRISE)

    ## Errors
    All errors follow standard HTTP status codes:
    - 400: Invalid request (validation error)
    - 401: Unauthorized (missing/invalid auth)
    - 403: Forbidden (quota exceeded or feature not available)
    - 404: Not found (job, profile, or submission not found)
    - 500: Server error

  version: 1.0.0
  contact:
    name: In–Midst–My–Life Support
    url: https://example.com/support
  license:
    name: MIT
servers:
  - url: http://localhost:3001
...
paths:
...
  /hunter/find-jobs:
    post:
      operationId: findJobs
      summary: Search and rank jobs by compatibility
      description: |
        Searches for jobs matching specified criteria and automatically
        ranks them by compatibility with user's profile.

        **Cost**: 1 search quota per request (FREE: 5/month, PRO: unlimited)

        **Ranking**: Results sorted by match score (0-1, higher better)
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSearchRequest'
      responses:
        '200':
          description: Jobs found and ranked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                  - pagination
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RankedJob'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Quota exceeded - no more searches available this month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaExceededError'
        '500':
          $ref: '#/components/responses/ServerError'
  /hunter/tailor-resume/{jobId}:
    post:
      operationId: tailorResume
      summary: Generate resume tailored to specific job
      description: |
        Generates a customized resume tailored to a specific job listing.
        Highlights relevant skills and experience for the target role.

        **Cost**: 1 resume tailoring quota (FREE: unavailable, PRO: 5/month, ENTERPRISE: unlimited)

        **Quality**: Returns confidence score (0-1) indicating how well the
        resume matches the job description.
      tags:
        - Documents
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: ID of the job listing to tailor resume for
          example: "job_12345"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentRequestOptions'
      responses:
        '200':
          description: Resume generated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TailoredResume'
        '400':
          description: Invalid request or unsupported format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Feature unavailable or quota exceeded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FeatureNotAvailableError'
                  - $ref: '#/components/schemas/QuotaExceededError'
        '404':
          description: Job or profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /hunter/generate-letter/{jobId}:
    post:
      operationId: generateCoverLetter
      summary: Generate customized cover letter
      description: |
        Generates a personalized cover letter for a specific job listing.
        Supports multiple templates and tones to match company culture.

        **Cost**: 1 cover letter quota (FREE: unavailable, PRO: 3/month, ENTERPRISE: unlimited)

        **Templates Available**:
        - professional: Traditional business format
        - creative: More personal, less formal
        - direct: Concise, action-focused
        - academic: Research/academic positions
      tags:
        - Documents
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: ID of the job listing
          example: "job_12345"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LetterRequestOptions'
      responses:
        '200':
          description: Cover letter generated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/GeneratedCoverLetter'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Feature unavailable for tier or quota exceeded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FeatureNotAvailableError'
                  - $ref: '#/components/schemas/QuotaExceededError'
        '404':
          description: Job or profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /hunter/submit-application/{jobId}:
    post:
      operationId: submitApplication
      summary: Submit job application
      description: |
        Submits a complete job application with tailored resume and
        cover letter. Can auto-submit to job boards (PRO+ feature).

        **Cost**: Varies by options:
        - Basic submission: Consumes 1 resume tailoring + 1 cover letter quota
        - Auto-submit: Also consumes 1 auto-apply quota (PRO: 5/month)

        **Process**:
        1. Generates or uses custom resume
        2. Generates or uses custom cover letter
        3. Submits to job board (if auto-submit enabled)
        4. Returns tracking ID and confirmation code
      tags:
        - Applications
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: ID of the job to apply for
          example: "job_12345"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSubmissionRequest'
      responses:
        '200':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ApplicationSubmissionResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Auto-submit unavailable or quota exceeded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FeatureNotAvailableError'
                  - $ref: '#/components/schemas/QuotaExceededError'
        '404':
          description: Job or profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    JobSearchRequest:
      type: object
      properties:
        keywords:
          type: array
          items:
            type: string
          description: Keywords to search for
          example: ["TypeScript", "Node.js", "AWS"]
        location:
          type: string
          description: Job location filter
          example: "Remote"
        minSalary:
          type: integer
          description: Minimum salary in USD
          example: 150000
        maxSalary:
          type: integer
          description: Maximum salary in USD
          example: 250000
        experienceLevel:
          type: string
          enum: ["entry", "mid", "senior", "lead"]
          description: Desired experience level
          example: "senior"
        limit:
          type: integer
          default: 20
          description: Maximum results to return
          example: 10
        includeRemote:
          type: boolean
          default: true
          description: Include remote roles
        featureFlags:
          type: object
          properties:
            prioritizeCompanyCulture:
              type: boolean
              default: false
            preferHigherScore:
              type: boolean
              default: true
    DocumentRequestOptions:
      type: object
      properties:
        format:
          type: string
          enum: ["pdf", "docx", "markdown"]
          default: "markdown"
          description: Desired output format
        highlightGaps:
          type: boolean
          default: true
          description: Highlight skill gaps and improvements
    LetterRequestOptions:
      type: object
      properties:
        template:
          type: string
          enum: ["professional", "creative", "direct", "academic"]
          default: "professional"
          description: Letter template/style
        tone:
          type: string
          enum: ["formal", "conversational", "enthusiastic"]
          default: "formal"
          description: Tone of voice
        includeSalutation:
          type: boolean
          default: true
        includeSignature:
          type: boolean
          default: true
...
    RankedJob:
      type: object
      required:
        - id
        - title
        - company
        - score
      properties:
        id:
          type: string
          description: Job listing ID
          example: "job_12345"
        title:
          type: string
          description: Job title
          example: "Senior Software Engineer"
        company:
          type: string
          description: Company name
          example: "TechCorp Inc"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Compatibility score with user profile
          example: 0.89
        location:
          type: string
          description: Job location
          example: "Remote"
        salary:
          type: object
          properties:
            min:
              type: integer
              example: 200000
            max:
              type: integer
              example: 280000
            currency:
              type: string
              default: "USD"
        description:
          type: string
          description: Job description excerpt
        url:
          type: string
          format: uri
          description: Link to job posting
        benefits:
          type: array
          items:
            type: string
          description: Highlighted perks or benefits
    TailoredResume:
      type: object
      required:
        - profileId
        - jobId
        - jobTitle
        - resume
        - metadata
      properties:
        profileId:
          type: string
        jobId:
          type: string
        jobTitle:
          type: string
        jobCompany:
          type: string
        resume:
          type: object
          properties:
            content:
              type: string
              description: Full resume content
            format:
              type: string
              enum: ["pdf", "docx", "markdown"]
            confidence:
              type: number
              format: float
              description: How well resume matches job (0-1)
              example: 0.85
            keywordMatches:
              type: array
              items:
                type: string
              description: Keywords from job found in resume
              example: ["TypeScript", "System Design", "AWS"]
            suggestedImprovements:
              type: array
              items:
                type: string
              description: Ways to improve resume match
        metadata:
          type: object
          properties:
            generatedAt:
              type: string
              format: date-time
            remainingTailorings:
              type: integer
              description: Tailorings remaining this month
              example: 3
            expiresAt:
              type: string
              format: date-time
              description: When quota resets
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 124
        pages:
          type: integer
          example: 7
...
    ErrorResponse:
      type: object
      required:
        - ok
        - error
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          description: Error code
          example: "NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "Profile not found"
        details:
          type: object
          additionalProperties:
            type: string
    ValidationError:
      type: object
      required:
        - ok
        - error
        - message
        - fields
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
          example:
            keywords: "Keywords array is required"
            minSalary: "Must be a positive number"
    QuotaExceededError:
      type: object
      required:
        - ok
        - error
        - message
        - context
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "QUOTA_EXCEEDED"
        message:
          type: string
        context:
          type: object
          properties:
            feature:
              type: string
              example: "hunter_job_searches"
            limit:
              type: integer
              example: 5
            used:
              type: integer
              example: 5
            remaining:
              type: integer
              example: 0
            resetsAt:
              type: string
              format: date-time
    FeatureNotAvailableError:
      type: object
      required:
        - ok
        - error
        - message
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "FEATURE_NOT_AVAILABLE"
        message:
          type: string
          example: "Feature 'cover_letter_generation' is not available in FREE tier"
        availableInTiers:
          type: array
          items:
            type: string
          example: ["PRO", "ENTERPRISE"]
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
                example: "UNAUTHORIZED"
              message:
                type: string
                example: "Authentication required"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
                example: "INTERNAL_ERROR"
              message:
                type: string
                example: "An unexpected error occurred"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
security:
  - BearerAuth: []
...